# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AfriToit - Gestion des logements",
        "description": "Documentation de l'API pour la gestion des logements",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/utilisateurs": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Récupérer tous les utilisateurs",
                "description": "Affiche la liste des utilisateurs.",
                "operationId": "27a03cbccef26b92bf6d80712f8b8ab8",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateurs"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Impossible de récupérer les utilisateurs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/utilisateurs/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Récupérer un utilisateur par ID",
                "description": "Affiche un utilisateur spécifique.",
                "operationId": "565dd7bc723d9a51b807156bd330fb2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateurs"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre à jour un utilisateur",
                "description": "Met à jour un utilisateur existant.",
                "operationId": "08fd8a9313115a94181f3651f05b6fa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Utilisateurs"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour"
                    },
                    "500": {
                        "description": "Impossible de mettre à jour l'utilisateur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Supprimer un utilisateur",
                "description": "Supprime un utilisateur.",
                "operationId": "c0fc30a3d234d962f29c73d25304de83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Utilisateur supprimé avec succès"
                    },
                    "500": {
                        "description": "Impossible de supprimer l'utilisateur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Utilisateurs": {
                "required": [
                    "idU",
                    "nomU",
                    "emailU"
                ],
                "properties": {
                    "idU": {
                        "description": "Identifiant unique de l'utilisateur",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nomU": {
                        "description": "Nom de l'utilisateur",
                        "type": "string"
                    },
                    "prenomU": {
                        "description": "Prénom de l'utilisateur",
                        "type": "string"
                    },
                    "emailU": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string",
                        "format": "email"
                    },
                    "telphoneU": {
                        "description": "Numéro de téléphone de l'utilisateur",
                        "type": "string"
                    },
                    "profilU": {
                        "description": "Profil de l'utilisateur",
                        "type": "string"
                    },
                    "statutU": {
                        "description": "Statut de l'utilisateur",
                        "type": "string"
                    },
                    "urlPhotoU": {
                        "description": "URL de la photo de l'utilisateur",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrer le token d'accès sous la forme Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}